[REF] link_tracker: reorganize python according to guidelines
This was not correctly done when splitting the module at a6290556bd151148241ec04a5129d26a8b0dc053.
Future commits will clean a bit the code in order to better understand how
this module works.

This commit is linked to task ID 1896681 and PR #27975. No functional change
should occur with this commit as it contains only renaming / move.