[FIX] product, point_of_sale, sale_management: Hide product sales tab
The sales tab is displayed on product form even if there are no modules
that need it are installed and no fields are displayed in it.

It has been introduced in this commit: https://github.com/odoo/odoo/commit/d9c9ed6cfbe028c308619a739ad350bc2d04be68
that we revert and fix by adding an xpath in the point_of_sale module
that display the sales tab when point_of sale_module is installed (even
if sale module is not installed.

Task ID - 1893455

closes odoo/odoo#28176